+++ frontend/packages/client/src/components/Community/CommunityEditorLinks/index.js
@@ -1,0 +1,2 @@
+export { default as CommunityEditorLinks } from './CommunityEditorLinks';
+export { CommunityLinksForm } from './CommunityEditorLinks';
